refactor: 거리 계산 시 소수점을 사용합니다.

+++ android/app/src/main/java/io/coursepick/coursepick/data/course/DefaultCourseRepository.kt
@@ -22,7 +22,7 @@ class DefaultCourseRepository(
                 mapCoordinate.longitude.value,
                 userCoordinate?.latitude?.value,
                 userCoordinate?.longitude?.value,
-                scope.meter.value,
+                scope.meter.value.toInt(),
             ).mapNotNull(CourseDto::toCourseOrNull)
 
     override suspend fun routeToCourse(
